Modified timeout judgment code in record_ir() of arduino_irrec.ino
This is required because some remote controller has a short gap
between the final data bit and a repeat leader (e.g.  Toshiba air
conditioner).  In the case, TIMEOUT_IR can not detect a repeat leader,
so we need to set timeout length as same as len_leader0 (or tab[1])
(refer analyze_pat()).